+++ CHANGELOG.md
@@ -4,6 +4,7 @@ All notable changes to this project will be documented in this file.
 
 ## [1.7.1-SNAPSHOT]
 
+- ObservableList.invalidate()
 - Dialog.toFront()
 - Node.whenVisible
 - ListCellFragment.onEdit